pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hw10-rag-api"
version = "0.1.0"
description = "Production-ready RAG API: FastAPI + SSE streaming, semantic cache, rate limiting, multi-provider fallback, cost tracking, Langfuse observability."
requires-python = ">=3.11"
readme = "README.md"
authors = [{ name = "MTeslenko" }]

dependencies = [
    # --- FastAPI stack
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",

    # --- LLM
    "openai>=1.50",         # used with OpenRouter via base_url
    "anthropic>=0.39",      # for Phase 2 LLM-as-judge + report interpretations

    # --- RAG
    "sentence-transformers>=3.2",
    "qdrant-client>=1.12",
    "langchain-text-splitters>=0.3",
    "tiktoken>=0.8",
    "pypdf>=5.0",

    # --- Storage
    "redis>=5.2",
    "httpx>=0.27",          # also used for Upstash Redis REST API
    "asyncpg>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",

    # --- Observability
    "langfuse>=2.50",
    "structlog>=24.4",

    # --- Utils
    "python-dotenv>=1.0",
    "tenacity>=9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.7",
]
report = [
    "matplotlib>=3.9",
    "pandas>=2.2",
    "weasyprint>=63",       # Windows: requires GTK runtime — see SETUP.md troubleshooting
    "jinja2>=3.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "W"]
ignore = ["E501"]
